Launches Printable onMessage
Working message passing to launch Printable from Citable.

For some reason it throws an error related to not being able to define
the background page when it is triggered from the Sheets UI. Odd. The
background page definitely gets the message and it seems to launch.

Need to pass the docId in the message.